Issue 935: fixes by waiting for bucket to be created before upload of blob
Commenting out this line (i.e. not creating the container) gives the KeyNotFoundException when the blob is big enough to require multiple parts. If only one part, it throws a org.jclouds.blobstore.ContainerNotFoundException.

Without the .get(), then for a small blob it works fine. But for a blob big enough to be multi-part then it throws the KeyNotFoundException. So something deeper is probably also going on here...