Added chef basics example
This example combines the compute-basics example with the Chef features
in jclouds to bootstrap the created nodes.

When using the SOLO action, the node will download the cookbooks from
the Opscode github repository and will run them without interacting with
a Chef Server.

When using the CHEF action, the node will be registered in the provided
Chef Server (can be a Community Chef, a Hosted chef or a Private Chef)
and will run the cookbooks accordingly.